backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
opencv-python
pillow
pytesseract
pandas
matplotlib
python-dotenv
